moved component to new dir
